[improvement] optimize isNaN (#3988)
