+++ src/utils/validate/date.ts
@@ -1,0 +1,7 @@
+import { isNaN } from './number';
+
+export function isDate(date: Date): boolean {
+  return (
+    Object.prototype.toString.call(date) === '[object Date]' && !isNaN(date.getTime())
+  );
+}
